requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cryptography==43.0.3
cymem==2.0.8
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
filetype==1.2.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langchain-together==0.2.0
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.136
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==2.0.2
olefile==0.47
openai==1.52.0
orjson==3.10.9
packaging==24.1
pandas==2.2.3
preshed==3.0.9
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pypdf==5.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.2
tika==2.6.0
tiktoken==0.8.0
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.0
unstructured-client==0.26.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.15.5
